requirements.txt
# Telegram Bot API
python-telegram-bot>=20.0

# Environment variables
python-dotenv>=1.0.0

# Google API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Spreadsheet handling
gspread>=5.10.0
pandas>=2.0.0

# Natural Language Processing
openai>=1.0.0

# Utilities
pytz>=2023.3